aiohttp/requirements.txt
json
sqlachemy
python-dotenv
aiodns==3.2.0
aiohttp==3.10.5
asyncpg==0.29.0
isort==5.13.2
